test(disaggregation): consolidate PD tests into test/srt and register in CI
