+++ test/srt/disaggregation/test_pd_auth.py
