+++ test/srt/run_suite.py
@@ -287,18 +287,30 @@ suites = {
         TestFile("test/srt/test_tokenizer_manager_event.py", 0.1),
         TestFile("test/srt/disaggregation/test_wrapper.py", 0.2, runner="pytest"),
         TestFile("test/srt/disaggregation/test_transport.py", 0.3, runner="pytest"),
-        TestFile("test/srt/disaggregation/test_conn.py", 0.5, runner="pytest"),
-        TestFile("test/srt/disaggregation/test_infra.py", 0.3, runner="pytest"),
         TestFile("test/srt/disaggregation/test_zmq_pull_notifier.py", 0.2, runner="pytest"),
         TestFile("test/srt/disaggregation/test_bootstrap_server.py", 0.5, runner="pytest"),
         TestFile("test/srt/disaggregation/test_pd_utils.py", 0.3, runner="pytest"),
-        TestFile("test/srt/disaggregation/test_scheduler_mixins.py", 0.5, runner="pytest"),
         TestFile(
             "test/srt/disaggregation/test_tokenizer_bootstrap_passthrough.py",
             0.3,
             runner="pytest",
         ),
-        TestFile("test/srt/disaggregation/test_pd_e2e_single_host.py", 1.0, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_mini_lb.py", 0.5, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_pd_prefill.py", 0.5, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_pd_decode.py", 0.5, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_pd_metrics.py", 0.3, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_pd_auth.py", 0.3, runner="pytest"),
+        TestFile("test/srt/disaggregation/test_pd_lifecycle.py", 0.5, runner="pytest"),
+        TestFile(
+            "python/sgl_jax/test/test_model_runner_kv_cache_mixin.py",
+            0.2,
+            runner="pytest",
+        ),
+        TestFile(
+            "python/sgl_jax/test/test_tp_worker_overlap_thread.py",
+            0.2,
+            runner="pytest",
+        ),
         TestFile("python/sgl_jax/test/test_compilation_manager.py", 1),
         TestFile("python/sgl_jax/test/test_kernel_utils.py", 1),
         TestFile("python/sgl_jax/test/speculative/test_spec_info.py", 0.2, runner="pytest"),
